Add PIX receivement status
Added a set of new files to project in order to handle the PIX receivement status. This includes a new client
class, validation rules, relevant types, and related testing methods.
This is to give users the ability to fetch and review the status of PIX receivements, increasing the functionality of the project.