Add getStatus method to CelcoinPIXReverse class
A new constant PIX_REVERSE_GET_STATUS_ENDPOINT was added to represent the getStatus API endpoint. Also, a new method called getStatus was created in the CelcoinPIXReverse class to get the status of a reversed PIX transaction. This change was required to interact with the getStatus API endpoint and retrieve the status of reversed PIX transactions via the SDK.