"Updated decimal precision and added PixWebhooks client"
In this commit, we have updated the decimal value in PixReverseCreate.php and PaymentInitRules.php. Instead of a precision of 2 decimal places, we are now only considering entire numbers.

Moreover, we added the CelcoinPixWebhooks client in Celcoin.php and its reference in CelcoinFacade.php. This allows the application to handle webhooks for Pix transactions.