Add functionality to reactivate and resend all pending Pix webhook messages
This commit adds a new feature in the CelcoinPixWebhooks client, allowing to reactivate and resend all pending Pix webhook messages. Implemented templates for two new types to be checked and to be handled respectively: PixReactivateAndResendAllPendingMessages and GenericWebhookErrorsTrait. Also, updated the tests accordingly. This new feature will facilitate addressing and managing of failed or pending webhook delivery attempts.