Add functionality to reactivate Pix events and resend specific messages
Introduced code that allows the reactivation of Pix events and resending of specified messages in a list. This functionality was added to the CelcoinPixWebhooks class and a new rule was created specifically for these actions. This is crucial for handling situations where an event needs to be reactivated and specific messages are required to be resent. Also, the related integration tests were added to ensure that the functionality works as intended.