Update Laravel workflow formatting and remove .env copy step
The formatting of the push and pull request branches, as well as the PHP version matrix in the Laravel GitHub workflow file has been updated for clarity and consistency. The .env copy step has been removed as it's unnecessary in our workflow, causing potential errors when it doesn't find the .env file. This change will improve the efficiency of our GitHub Actions workflows.